Split reflexion worker responsibilities
Keep ReflexionWorker focused on lifecycle and queue orchestration while moving reflection persistence and procedural memory conversion into dedicated infrastructure services.

Constraint: issue #599 requires sub-500-line units and reflexion regression preservation

Rejected: rewriting public worker callers | unnecessary for a behavior-preserving split

Confidence: high

Scope-risk: moderate

Directive: keep ReflexionWorker as orchestration only; put persistence and conversion changes in the extracted services

Tested: npm test -- packages/memento-core/src/infrastructure/reflexion-worker.spec.ts; npm run type-check; npm run lint; git diff --check; graphify rebuild

Not-tested: full npm test suite
